ResNet50 pre-trained unfroze pre-trained model change step per epoch
Final score of 64702. I unfroze the last 10 layers in pre-trained model but this reduced the generalization score as it caused the model of overfit given the small size data, so rolled back to total freeze and adjusted the steps_per_epoch inline with best practice which resulted in the best score. with early stopping to prevent overtraining and making void needed to change other hyper parameters e.g. batch size, epochs, learning rate (used a auto learning alogorithm).